story-7 refactor podcast list
