+++ Silkrode/app/src/main/java/personal/ivan/silkrode/navigation/podcast/view/PodcastActivity.kt
@@ -1,12 +1,8 @@
 package personal.ivan.silkrode.navigation.podcast.view
 
 import android.os.Bundle
-import android.view.View
 import androidx.activity.viewModels
-import androidx.lifecycle.Observer
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import dagger.android.support.DaggerAppCompatActivity
-import personal.ivan.silkrode.R
 import personal.ivan.silkrode.databinding.ActivityPodcastBinding
 import personal.ivan.silkrode.di.AppViewModelFactory
 import personal.ivan.silkrode.navigation.podcast.viewmodel.PodcastViewModel
@@ -39,47 +35,6 @@ class PodcastActivity : DaggerAppCompatActivity() {
      */
     private fun observeLiveData() {
         mViewModel.apply {
-            // API status - loading
-            apiLoading.observe(
-                this@PodcastActivity,
-                Observer { switchLoadingProgress(enable = it) })
-
-            // API status - fail
-            apiFail.observe(
-                this@PodcastActivity,
-                Observer { showApiError() })
-
-            // Toolbar title text
-            toolbarTitle.observe(
-                this@PodcastActivity,
-                Observer { updateToolbarTitle(it) })
         }
-    }
-
-    /* ------------------------------ UI */
-
-    /**
-     * Update Toolbar title
-     */
-    private fun updateToolbarTitle(title: String) {
-        mBinding.toolbar.title = title
-    }
-
-    /**
-     * Show or hide loading progress
-     */
-    private fun switchLoadingProgress(enable: Boolean) {
-        mBinding.progressBarLoading.visibility = if (enable) View.VISIBLE else View.GONE
-    }
-
-    /**
-     * Show API error
-     */
-    private fun showApiError() {
-        MaterialAlertDialogBuilder(this)
-            .setTitle(R.string.alert_title)
-            .setMessage(R.string.alert_message)
-            .setPositiveButton(R.string.label_ok, null)
-            .show()
     }
 }
